move special cases transformation to controlTypes
