+++ source/IAccessibleHandler/__init__.py
@@ -296,7 +296,7 @@ def calculateNvdaRole(IARole: int, IAStates: int) -> Role:
 	"""
 	role = IAccessibleRolesToNVDARoles.get(IARole, Role.UNKNOWN)
 	states = _getStatesSetFromIAccessibleStates(IAStates)
-	role, states = transformRoleStates(role, states)
+	role, states = controlTypes.transformRoleStates(role, states)
 	return role
 
 
@@ -305,7 +305,7 @@ def calculateNvdaStates(IARole: int, IAStates: int) -> Set[State]:
 	"""
 	role = IAccessibleRolesToNVDARoles.get(IARole, Role.UNKNOWN)
 	states = _getStatesSetFromIAccessibleStates(IAStates)
-	role, states = transformRoleStates(role, states)
+	role, states = controlTypes.transformRoleStates(role, states)
 	return states
 
 
@@ -318,31 +318,6 @@ def NVDARoleFromAttr(accRole: Optional[str]) -> Role:
 	else:
 		accRole = accRole.lower()
 	return IAccessibleRolesToNVDARoles.get(accRole, controlTypes.Role.UNKNOWN)
-
-
-def transformRoleStates(role: Role, states: Set[State]) -> Tuple[Role, Set[State]]:
-	""" Map NVDA Role-State combinations to adjusted NVDA Role-State combinations.
-	Some combinations of roles and states may be better represented with some alternative combination.
-	As an example:
-	Role.PROGRESSBAR with State.INDETERMINATE should be represented by only the Role.BUSY_INDICATOR, with
-	no State.INDETERMINATE, or State.HALFCHECKED.
-	@param role: NVDA Role to consider
-	@param states: NVDA States to consider
-	@return: A tuple with the new Role and modified States set.
-	"""
-	if(
-		role in [Role.PROGRESSBAR, Role.BUSY_INDICATOR]
-		and states.intersection({State.INDETERMINATE, State.HALFCHECKED})
-	):
-		# Don't report indeterminate progress bars as "half-checked"
-		# L{State.HALFCHECKED} maps from oleacc.STATE_SYSTEM_MIXED,
-		# which has the same value as oleacc.STATE_SYSTEM_INDETERMINATE.
-		# L{State.INDETERMINATE} is not mapped directly from any IA or IA2 state.
-		# A progress bar that can not convey progress, only activity, is a busy indicator.
-		states.discard(State.HALFCHECKED)
-		states.add(State.INDETERMINATE)
-		return Role.BUSY_INDICATOR, states
-	return role, states
 
 
 def normalizeIAccessible(pacc, childID=0):
